# api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.AuthResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        example: "2026-03-16T18:00:00Z"
        type: string
      refreshToken:
        type: string
      user:
        $ref: "#/definitions/presenter.UserResponse"
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        example: validation error
        type: string
    type: object
  handlers.ForgotPasswordRequest:
    properties:
      email:
        example: jane@example.com
        type: string
    required:
      - email
    type: object
...
  handlers.LoginChallengeResponse:
    properties:
      email:
        type: string
      message:
        type: string
      otpExpiresAt:
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      email:
        example: jane@example.com
        type: string
      password:
        example: supersecret
        type: string
    required:
      - email
      - password
    type: object
  handlers.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
      - refreshToken
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        example: jane@example.com
        type: string
      name:
        example: Jane Doe
        type: string
      password:
        example: supersecret
        minLength: 8
        type: string
      preferredLocale:
        enum:
          - fr
          - en
        example: fr
        type: string
    required:
      - email
      - name
      - password
    type: object
  handlers.ResetPasswordRequest:
    properties:
      email:
        example: jane@example.com
        type: string
      newPassword:
        example: newsecret123
        minLength: 8
        type: string
      otp:
        example: "123456"
        type: string
    required:
      - email
      - newPassword
      - otp
    type: object
...
  handlers.VerifyLoginOTPRequest:
    properties:
      email:
        example: jane@example.com
        type: string
      otp:
        example: "123456"
        type: string
    required:
      - email
      - otp
    type: object
...
  presenter.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      mustChangePassword:
        type: boolean
      name:
        type: string
      preferredLocale:
        allOf:
          - $ref: "#/definitions/user.Locale"
        enum:
          - fr
          - en
        example: fr
      profilePhotoFileId:
        type: string
      profilePhotoUrl:
        type: string
      role:
        allOf:
          - $ref: "#/definitions/user.Role"
        enum:
          - admin
          - user
        example: user
      updatedAt:
        type: string
    type: object
  user.Locale:
    enum:
      - fr
      - en
    type: string
    x-enum-varnames:
      - LocaleFR
      - LocaleEN
  user.Role:
    enum:
      - admin
      - user
    type: string
    x-enum-varnames:
      - RoleAdmin
      - RoleUser
info:
  contact: {}
  description:
    API server built with Gin, Gorm, Postgres, JWT auth, refresh tokens,
    and Swaggo.
  title: Boilerplate Go Vue API
  version: "1.0"
paths:
  /api/v1/auth/forgot-password:
    post:
      consumes:
        - application/json
      description: Sends a password reset OTP by email if the account exists.
      parameters:
        - description: Forgot password payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/handlers.ForgotPasswordRequest"
      produces:
        - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      summary: Forgot password
      tags:
        - auth
  /api/v1/auth/login:
    post:
      consumes:
        - application/json
      description: Authenticates a user and sends a 2FA OTP by email.
      parameters:
        - description: Login payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/handlers.LoginRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handlers.LoginChallengeResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      summary: Login
      tags:
        - auth
  /api/v1/auth/logout:
    post:
      consumes:
        - application/json
      description: Revokes a refresh token.
      parameters:
        - description: Logout payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/handlers.RefreshRequest"
      produces:
        - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      summary: Logout
      tags:
        - auth
  /api/v1/auth/refresh:
    post:
      consumes:
        - application/json
      description: Rotates the refresh token and returns a new token pair.
      parameters:
        - description: Refresh payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/handlers.RefreshRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handlers.AuthResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      summary: Refresh tokens
      tags:
        - auth
  /api/v1/auth/register:
    post:
      consumes:
        - application/json
      description: Creates a user account. Role defaults to `user`.
      parameters:
        - description: Register payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/handlers.RegisterRequest"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/presenter.UserResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "409":
          description: Conflict
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      summary: Register a user
      tags:
        - auth
  /api/v1/auth/reset-password:
    post:
      consumes:
        - application/json
      description: Resets the password using the OTP sent by email.
      parameters:
        - description: Reset password payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/handlers.ResetPasswordRequest"
      produces:
        - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      summary: Reset password
      tags:
        - auth
  /api/v1/auth/verify-otp:
    post:
      consumes:
        - application/json
      description: Verifies the login 2FA OTP and returns JWT tokens.
      parameters:
        - description: Verify login OTP payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/handlers.VerifyLoginOTPRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handlers.AuthResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/handlers.ErrorResponse"
      summary: Verify email OTP
      tags:
        - auth
...
securityDefinitions:
  BearerAuth:
    description: Use `Bearer <token>`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"